Add files via upload
This is the code that I'll be using on the first two pis to be deployed.  Obviously the file names are bulky, but I'd like to know when strikes are happening.  I'm going to get in contact with IT to see if we connect each pi to bluezone and then view their desktops remotely.